Object Array Creation & user input
This Java code defines a Person class with name and age attributes and a display() method. The main method takes user input for the number of persons, then for each person's name and age, creates Person objects, stores them in an array, and finally displays the information of all persons.  It uses a Scanner to get input from the user.